requirements.txt
hydralit_components==1.0.10
openpyxl<3.0.10
pandas==1.4.3
Pillow==9.2.0
xlrd<2.0.1
regex==2022.7.25
streamlit==1.11.1
streamlit_option_menu==0.3.2
xlwings==0.27.11
